Copy input data to local SSD before reading.
Files are copied from the network drive to local SSD, to speed up the reading the file. The routine still preforms processing line-by-line, but now the entire row is read at the beginning and stored into an array (instead of reading each individual pixel separately).

Added are some other functionalities:
- specify at input whether the file should be copied (TODO: if not, read directly from network drive (slower));
- switch parallel on/off